Refactor routes into files and cleanup imports.
