+++ admin/admin_context.py
@@ -1,3 +1,5 @@
+from admin import admin_config
+
 
 # Useful context to be passed around the all the other admin file
 class Context:
@@ -9,3 +11,12 @@ class Context:
         self.server_host = server_host
         self.server_port = server_port
         self.server_user = server_user
+
+    @classmethod
+    def load_from_db(cls, league_name):
+        configs = admin_config.get_leagues()
+        return Context(league_name,
+                       configs[admin_config.BOT_COMMAND],
+                       configs[admin_config.SERVER_HOST],
+                       configs[admin_config.SERVER_PORT],
+                       configs[admin_config.SERVER_USER])
