+++ admin/deploy.py
@@ -1,7 +1,7 @@
 import sftp
 import os
 import sqlite3
-from admin_context import Context
+from admin.admin_context import Context
 
 server_folders = ['backend']
